Move `Queries` under the `client` package
