+++ server/src/main/java/org/spine3/server/stand/Stand.java
@@ -26,9 +26,9 @@ import com.google.common.util.concurrent.MoreExecutors;
 import com.google.protobuf.Any;
 import com.google.protobuf.Message;
 import io.grpc.stub.StreamObserver;
-import org.spine3.base.Queries;
 import org.spine3.base.Responses;
 import org.spine3.base.Version;
+import org.spine3.client.Queries;
 import org.spine3.client.Query;
 import org.spine3.client.QueryResponse;
 import org.spine3.client.Subscription;
