+++ server/src/test/java/org/spine3/server/SubscriptionServiceShould.java
@@ -23,8 +23,8 @@ package org.spine3.server;
 import com.google.protobuf.Message;
 import io.grpc.stub.StreamObserver;
 import org.junit.Test;
-import org.spine3.base.Queries;
 import org.spine3.base.Response;
+import org.spine3.client.Queries;
 import org.spine3.client.Subscription;
 import org.spine3.client.SubscriptionUpdate;
 import org.spine3.client.Target;
