+++ server/src/test/java/org/spine3/server/stand/SubscriptionRecordShould.java
@@ -23,7 +23,7 @@ package org.spine3.server.stand;
 import com.google.protobuf.Any;
 import com.google.protobuf.Message;
 import org.junit.Test;
-import org.spine3.base.Queries;
+import org.spine3.client.Queries;
 import org.spine3.client.Subscription;
 import org.spine3.client.Target;
 import org.spine3.protobuf.AnyPacker;
